[#38] refactor: Rename back-end functions in services folder
