+++ .gitignore
@@ -14,5 +14,5 @@ yarn.lock
 # macOS
 .DS_Store
 
-# API KEYS
-api/keys.tsx
+# API KEYS (나중에 공개할지 말지 정하기)
+api/keys.ts
